fixed various bugs, lots of refactoring

+++ Adria/Graphics/GfxTexture.h
@@ -65,7 +65,7 @@ namespace adria
 	class GfxTexture
 	{
 	public:
-		GfxTexture(GfxDevice* gfx, GfxTextureDesc const& desc, GfxTextureInitialData* initial_data = nullptr, size_t data_count = -1);
+		GfxTexture(GfxDevice* gfx, GfxTextureDesc const& desc, GfxTextureInitialData* initial_data = nullptr, uint32 data_count = -1);
 		//constructor used by swapchain for creating backbuffer texture
 		GfxTexture(GfxDevice* gfx, GfxTextureDesc const& desc, ID3D12Resource* backbuffer);
